requirements.txt
boto3==1.26.22
botocore==1.29.22
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
contourpy==1.0.7
cryptography==40.0.1
cssselect==1.2.0
cycler==0.11.0
dnspython==2.3.0
docopt==0.6.2
fonttools==4.38.0
html5-parser==0.4.10
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.3
mtools==1.7.2
mut==0.9.2
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psutil==5.9.4
pycparser==2.21
pymongo==4.3.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyspellchecker==0.7.2
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
